% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\docType{data}
\name{koch}
\alias{koch}
\title{Koch's snowflake}
\format{An object of class \code{list} of length 4.}
\usage{
Recommended k > 6
}
\description{
Koch's snowflake
}
\keyword{datasets}
